Migrate C-HashMap to submodule
Following upstream, we're now also no longer copy-pasting C-HashMap
source files directly.

Like FAudio, this is a C99 library, so it originally required changes
to at least work with C90-with-extensions. Now that we're moving to
C99, it doesn't need any changes anymore and I can just take the
upstream version as-is.