Change explicit C-HashMap function includes to #include inside extern
C-HashMap is a C library which doesn't guard its header file with an
extern "C", so I tried this construct originally to include it:

    extern "C"
    {
    #include <map.h>
    }

This worked on my GCC, but it failed on my Visual Studio. After some
debugging and looking online, the answer I'm pretty sure I found online
was that MSVC didn't support #include inside an extern block, and that
it was nonstandard and undefined behavior that's just asking for
trouble anyway.

However, during the review for this PR, Misa and I spent about 1 hour
and 15 minutes trying to confirm that it was in fact forbidden, and
found nothing to confirm it.

As it turns out, the actual problem seems to be that MSVC prioritizes
some map.h of its own, and that's why the #include seemed to be
ignored. It looks like `#include <../c-hashmap/map.h>` also works, and
I want to try an even better fix in the next commit, but I also wanna
see what the Windows CI says, hence this commit.