Make max_check_buffer in loadtext_strings_plural not malloced
I couldn't really remember what the reasoning was for this conditional
mallocing, the best I still can think of is that I wanted to not waste
3.2 KB of memory on an internal testing feature every time strings are
loaded when 99.99% of people will be regular players who won't use that
feature. That, and the code possibly evolved a little bit while I was
implementing it and never considered that this part of the design could
now be simplified. Maybe I wanted to make the buffer length dynamic at
first too, for example.

3201 bytes is not that much to allocate on the stack 100% of the time,
and loading strings is not a hot path (it's done only once upon startup
and maybe once more during events like changing the language).

But why make the buffer in loadtext_strings_plural anyway? It's only
passed to max_check_string_plural and used inside that function... that
function could just make a buffer itself. Again, this is probably an
artifact of an earlier design stage that could now be simplified.

https://youtu.be/MSaCmn9S3pI?t=4206