Add UNUSED for unused arguments in callback_free_map_value
https://youtu.be/fHJyTwn35Mk?t=339
(sort of, the audio in this part of the VOD is unfortunately broken -
my voice is duplicated and Misa's voice is missing)

Basically, with strict enough warnings enabled, compilers can warn for
functions having unused arguments. When you create a callback function
for a library however, you don't have control over the arguments, and
you may not use all of them. That's what UNUSED is for, to indicate the
variable is unused on purpose and to silence the warning.