Migrate plural strings from snprintf to VFormat
This changes how loc::gettext_plural_fill works: it now takes varargs
and passes them on to VFormat, so this is no longer limited to cases
where there's only one argument to be filled in (thus far all cases).
The first vararg is the one that determines the plural string, but that
restriction doesn't matter for VFormat - you may only need to reorder
your arguments while leaving the string itself the same, and the
argument doesn't even have to occur in the string itself.

This also changes max checking slightly. Plural strings now have a
"var" attribute which indicates the specific variable that will be
tested. Previously, strings containing %d would be tested for the
specific indicated expected number (like 1000) while strings containing
%s would be tested for each wordy number until the expected number
(Seven might make a string too long while Ten doesn't). Now that it's
a bit harder to know whether a string will be filled with wordy
numbers, instead expected numbers of 100 and below will be tested from
0 to N (even if it's numeric, we can easily take this number of tests)
and expected numbers above 100 will be done with only that number (you
shouldn't make that wordy).

TODO: go through lang/README.txt and update it for VFormat. And also
update the Excel sheet for "var" in strings_plural.xml.