Finish Dutch cutscenes translations
I've translated everything that hadn't yet been, and I've also improved
the existing translations.

One choice I made is to use "de teleport" for "the teleporter"
everywhere. I guess "de teleporter" sounded vaguely normal-but-off to
me since I can't really remember hearing references to a teleporter as
an object in Dutch (only really as a verb) and it's hard to find any.
And I've known this game for a while so it might've started to get into
my own Dutch thoughts as a kind of loan word.

I did find usage of "de teleport" in a Dutch-language game from 1995
and its documentation though, so it seems fine to go with it. It sounds
better than anything else I can think of.