Redesign time trials menu once more for a two-column layout
This has come up in the speedrunning server. The time trials menu used
to be a bit too tight to even add centiseconds to the best time
display, but I redesigned the menu for localization, so now there would
be enough space for it. Except, the redesign was to give more breathing
room for languages which need longer words here, and adding
centiseconds would kind of nullify that extra space.

So, this is another redesign: the left column (with the records) stays
the same, the par time moves to the right, and the best rank moves
down. The best rank now sticks out on the bottom, but it still looks
good. The menu has also been padded a bit more, and there's still even
more space for languages and centiseconds than there has ever been.