Prevent F12 hotkey (reload language files) from being held down
You could hold down F12 and it would start repeatedly reloading
language files. This is now fixed so it only works once until you lift
the key and press it again. It's a debug key so it was pretty
inconsequential, but it's an easy fix.

https://youtu.be/tXhhKA5RtDA?t=1924
(I didn't hear the repeated coin sounds at the time because it wasn't
passed via the voice chat, but it became clear soon enough lol)