Fix translation statistics sometimes being right-aligned (on Windows)
A bit of history: In a related manner earlier on, it turned out VFormat
mysteriously broke on Windows specifically, and failed to fill in
many placeholders. The reason turned out to be that fallback
implementations for some SDL_ replacements to stdlib functions we
have been using, didn't follow the C standard... These fallback
implementations are used on Windows for some reason. I did test VFormat
on Windows, but that was with the standard strchr, Microsoft's
implementation of which works fine. It's just that SDL falls back to
its own implementation on Windows. So I made my mark on SDL code and
fixed it: https://github.com/libsdl-org/SDL/pull/5887

Fast-forward to here: https://youtu.be/tXhhKA5RtDA?t=4756

https://cdn.discordapp.com/attachments/380088789802024960/993687562214440970/unknown.png
https://cdn.discordapp.com/attachments/380088789802024960/993687670431690782/unknown.png

It turned out that the fallback implementation of SDL_snprintf is also
incomplete. Specifically, the format string "%-32s" is supposed to
left-align a string and pad with spaces on the right, but in SDL's
fallback implementation, it right-aligns it and pads with spaces on the
left. This time I'm just gonna change the code itself - it's simpler to
just have multiple prints instead of this format string - the reason
why I made it a printf string is probably because I was originally also
planning to add a percentage column before deciding not to.

Also, the translation statistics screen included the file
"roomname_special.xml", which should've been "roomnames_special.xml",
so that has been fixed too.