Add fallback instructions for unknown companion in Intermission 1
Intermission 1 has instructions at the top telling you how the other
crewmate follows you when you are/aren't standing on the floor/ceiling.
These instructions refer to the specific crewmate (by name and he/she).
But what if you go into Intermission 1 with an invalid "last rescued"
crewmate?

It used to be that, due to the way these messages were glued together
line-by-line, you'd get a malformed message like "When you're NOT
standing on stop and wait for you". This line-by-line work is
practically impossible to carry over for localization, so instead I
just made it default to Victoria instead.

As for what's changing now, to quote Misa on the VVVVVV Discord server:

> imo the default should instead just default to the crewmate being
> referred to as `the crewmate`
> maybe gamestate 12 should also check `game.lastsaved` and go
> 2: `he is`, 3: `he is`, 4: `he is`, 5: `she is` and for all other
> values use `they are`
> having custom levels in mind here

This is now done, except I found "the crewmate will try to walk to you"
so unnatural, even for a fallback, that I made it "your companion"
instead.