Draw outlines for colored characters correctly
If we try to draw a colored character, and the color is fully black,
it means we're drawing the outline.

Left todo: make colored characters fade out properly when drawn with
a different alpha (like when fading out the [Press ENTER to return to
editor]). This probably requires BlitSurfaceDarkened, but I haven't
fully looked into that yet.