Revert "Try adding line underneath imebuffer"
This reverts commit c6800ce9374b4670dd324376990974a0526bbd61.

Probably best to make this PR start simple, with just more transparent
text, and no underlines for the selected part. That can always be done
in a separate PR, but what I'm thinking right now is a font:: function
to draw an underline under a given segment of a given text at a certain
x,y position.