Add information strip to window title bar
This displays total game time and total number of deaths, and the time
and number of deaths since getting the last checkpoint.

Originally I wanted this to be something underneath the regular game
screen, so the game's resolution would be something like 320x250 or
something, but I quickly gave up trying to change the game's resolution
because I didn't know what to change and what to leave, so it felt like
a really bad idea. I also didn't get it working quickly enough.
Luckily the title bar works too.